+++ Project/Source Files/Services/Models/Recipe.swift
@@ -21,19 +21,3 @@ extension Recipe: Codable {
         case recipeURL = "href"
     }
 }
-
-struct Recipes {
-    let title: String
-    let version: Float
-    let recipesURL: String
-    let recipes: [Recipe]
-}
-
-extension Recipes: Codable {
-    enum CodingKeys: String, CodingKey {
-        case title = "title"
-        case version = "version"
-        case recipesURL = "href"
-        case recipes = "results"
-    }
-}
